backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
